requirements.txt
Django
gunicorn
Markdown
django-filter
djangorestframework
whitenoise
Python==3.10.6
pyscopg2
pyjwt
django-cors-headers